fill_server/requirements.txt
fastapi==0.71.0
uvicorn==0.16.0
python-dotenv==0.19.2
httpx==0.21.3
python-jose==3.3.0
